+++ src/msg_utils.h
@@ -16,7 +16,7 @@ namespace MSG_Utils {
     void    ledNotification();
     void    deleteFile(String typeOfFile);
     void    saveNewMessage(String typeMessage, String station, String newMessage);
-    void    sendMessage(int typeOfMessage, String station, String textMessage);
+    void    sendMessage(uint8_t typeOfMessage, String station, String textMessage);
     void    processOutputBuffer();
     void    checkReceivedMessage(ReceivedLoRaPacket packetReceived);
